+++ luatest/utils.lua
@@ -1,3 +1,4 @@
+local digest = require('digest')
 local fun = require('fun')
 local yaml = require('yaml')
 
@@ -146,6 +147,12 @@ function utils.get_fn_location(fn)
     local fn_details = debug.getinfo(fn)
     local fn_source = fn_details.source:split('/')
     return ('%s:%s'):format(fn_source[#fn_source], fn_details.linedefined)
+end
+
+function utils.generate_id(length, urlsafe)
+    if not length then length = 9 end
+    if urlsafe == nil then urlsafe = true end
+    return digest.base64_encode(digest.urandom(length), {urlsafe = urlsafe})
 end
 
 return utils